Improve key lookup and key_ranges.
Redefine how key-ranges work, and allow new registrations to
over-ride old registrations.
So you can define a range, then redefine individual keys.

Each entry in the key map and be the start of a range, in which case
all matching keys up to the next entry are recongnised, or not the
start of a range, in which case an exact match is required.

The same key can be recorded as a non-range and a range, in that order.
The first entry applies to an exact match, the second to the following range.
This allows for open-ended ranges.

Signed-off-by: NeilBrown <neil@brown.name>