Introduce 'pane_with_cursor'.
Some movement commands need to know about the current cursor location.
emacs_move() was using a kludge to find this - it may not keep working.
So introduce pane_with_cursor() to find the cursor pane.
Use this also to draw the cursor.

Removing this kuldge will allow point_pane to be revised.

Signed-off-by: NeilBrown <neil@brown.name>