Allow things other than panes to own a point.
A 'point' is where changes happen and changes that happen in
the background (e.g. when data arrives from socket) might
not have a pane, but still want a point.

So allow anything to own a point.  That means cmd_info doesn't need
point_pane, just a pointer to the point.

Signed-off-by: NeilBrown <neil@brown.name>