point_notify_change needs a range.
If a change happens at a mark to be notified, then there is no
guarantee that the mark is before the change point, it maybe be
afterwards in the list.  Or a mark that should be treated as after the
change could be listed before.

So we need to notify all marks that are at either of of a change.  So
pass a range as mark (start) and point (end) and notify anything
possibly relevant.

Signed-off-by: NeilBrown <neil@brown.name>