render_hex: make sure position numbers stay correct.
As each line shows the position in the file, we need to
update when there is any change before the end of the display,
not just after the start of the display.

Also if the start of the display is ever not a multiple of 16, it needs to be re-adjusted.

Signed-off-by: NeilBrown <neil@brown.name>