Centralise reseting of ->focus
If ->focus is NULL before calling refresh, set it to
an  arbitrary child.  Ideally other code will set it more proactively,
but this ensures it doesn't get lost, and ensures it all in one place.

Also make sure that when a pane loses focus it gets refreshed to ensure
focus gets updated.

Signed-off-by: NeilBrown <neil@brown.name>