Add render-lines.
render-lines renders a pane given a 'render-lines' function on the
underlying document.  It handles charact-wrapping or line-truncation
and side-shifting.
It determines how many lines need to be rendered to fill the pane,
and handles movement related to the pane (page up/down, line up/down).

This should make it easier to write individual render functions.

Signed-off-by: NeilBrown <neil@brown.name>