Add render-format.
This renders one document element per line using a format string
and attributes to create content, and using render-lines to display
on the pane.

This is particularly useful for dir listing and the document list.

This will replace render-dir.

This doesn't quite work right yet  - in particular lines up/down is wrong.

Signed-off-by: NeilBrown <neil@brown.name>