Change meaning for return value for "Refresh"
Allowing "Refresh" to return 0 was always wrong.  Returning zero
should mean no command as found.
Allowing it to return non-zero to mean "apply changes to children"
is now unused, so discard that.

Now all Refresh functions that do something return 1.  So if no such
function was found, call pane_check_size() to ensure that size matches
parent, as that is a basic requirement of simple panes.

Signed-off-by: NeilBrown <neil@brown.name>