render-lines: make sure cursor appears in empty file.
when point is at the end of a line it is also at the start of the next
line, and this can cause confusion.

Resolve by preferring the start of the next line (even if it is the
non-existent line at end of file) unless the end of the previous line
in not in the first column (e.g. no newline at end-of-file).

Signed-off-by: NeilBrown <neil@brown.name>