Remove the document list and use a list of panes instead.
The 'null display' (ed->root.focus) is now the parent of multiple
panes, each of which owns a document (->data) with a backlink
(doc->home->data == doc).

Now in many places we pass around this pane instead of a point when we
just want to identify a document and not a location in the document.
This means we don't need to allocate points so much, or fake up
temporary ones.

doc_attach() must be used at document creation to attach a document to
the tree.

So
 - "*Documents*" watches this list of siblings rather than the original list.
 - commands can be sent to these panes using key_handle_focus.  Some
   doc commands can be sent to any pane.
 - doc creation returns the pane of the document in ci->focus
   rather than a fake point
     e.g. same_file, load_file

 - pane_attach() takes a pane, not a point.
 - doc_open no-longer needs a gross hack
 - editor_choose_doc returns a pane, and so doesn't need a point.
 - popup holds the doc pane, not a point.

etc.

Signed-off-by: NeilBrown <neil@brown.name>