Initial python support.
Add a new loadable module 'lang-python' which provides a
"python-load" command which will run a python script.

This has access to panes and marks and can send and receive commands...
At least it can in principle - only limited testing so far.

I think it is a good starting point for python support.

Signed-off-by: NeilBrown <neil@brown.name>