Move the authoritative 'point' out of the pane and into the doc_data.
This seems a more sensible location.
The point still exists in the pane as it is very useful, but this is
now just a 'borrowed' pointer.

This required changing doc_attr to only examine a single pane.
Probably makes sense for now.

Next patch will make changes to how points in panes are managed.

Signed-off-by: NeilBrown <neil@brown.name>