Start changing rules about ->focus and ->home
I'm still thinking this through....

'home' must always be the pane where the command was found.
'focus' should be where the action happens.

Previously 'home' would always end up being a parent of 'focus' but
sometimes, particularly calls from popup to target, we want something else.

So allow the caller to set 'focus' to themselves and 'home' to where
to start looking for the command.
If either is home is not set, focuses it moved out to a leaf first.

This still seems a bit clumsy...

Signed-off-by: NeilBrown <neil@brown.name>