Revisions to pane_refresh
Instead of passing a pointer to a struct which is then copied and
otherwise (largely) unused, just pass the struct by-value.

Also change the setting of some fields that isn't needed as they
cannot have been changed.

And never use a ci again after it has been passed to a command.

Signed-off-by: NeilBrown <neil@brown.name>