python: bind python Pane more tightly to edlib pane
When python allocates a pane, the matching Pane is buond to it and
always used in python code (previously we would allocate various separate Panes).
This makes is possible to subclass Pane and when we get a pane back from
edlib, we get the whole subclassed Pane.

I should do a similar think with Marks....

Signed-off-by: NeilBrown <neil@brown.name>