mark: forward/backward must prefer ordering of refs
refs to a doc can be ordered - even when they are the same.
The end of one text-chunk is the same as the beginning of the next,
but they must never occur in the reverse order in the list of marks,
else confusion results.

So when we move one mark over another on noticing they are the 'same',
we must update the 'ref' too to ensure ordering is preserved.

Someones we move a mark over another when not the same.  In that case,
the ref will be updated correctly soon so changing it early is not a
problem.

Signed-off-by: NeilBrown <neil@brown.name>