python: Allow mark to be created with Mark()
A pane can be passed in which case a new mark will be created
at the start of the attached document - with given view number.
If no view number, then an anon mark is created.

Alternately a mark can be given and it will be duplicated - same
type is used.

Signed-off-by: NeilBrown <neil@brown.name>