Python/marks: discard data access, improve lifetime.
Instead of a 'data' accessor, let marks be subtyped.
This means that marks created by Python have fixed python
instances.  When the instance is freed, the mark is unlinked.

'vmarks' persist beyond the last reference from python and need to be
explicitly released. If you don't release all vmarks when a doc is
destroyed ... bad things happen.

Signed-off-by: NeilBrown <neil@brown.name>