python: make 'home' the primary pane.
When a command is called with two panes, the "Home" is asked to act
in the context of the "focus". So it makes sense to make it look
like a method call on the "Home" pane.

When only one pane is given, it makes sense for it to be both "home"
and "focus".  So if we switch now, and copy home to focus when the later is
not set, everything works as expected.  I hope.

Signed-off-by: NeilBrown <neil@brown.name>