doc: make home panes distinct from reference pane.
A document is always accessed through a pane.  There have long been to
sorts of panes, but they were confusingly similar at times.  Now they
are more different.

A "home" pane is created when the document is created.  The ->data of
the home pane is the document (struct doc *).  This pane should be
registered with "docs" so it can be found later.

A "reference" pane is created whenever a document is to be displayed.
->data is a "struct doc_data" which contains a "point" and a pointer
to the home pane.

->handle for the home pane contains the doc-specific handling.
->handle for a reference pane is doc_handle, and it will defer
to home->handle for some functionality.

doc_destroy() is now a real function, just a method call.  It can only
be called on the home pane.  A method call is still used on reference pane.

There are still some messy bits, particularly around document name handling,
and the passing of 'point' to the home document methods.

Signed-off-by: NeilBrown <neil@brown.name>