doc: create message-bus via documents.
One pane viewing a document can not listen for messages from other
panes viewing the same document.

This will be used for a presentation mode to move to a different page
in response to an action in a "presenter" view.
Maybe it will be used for other things one day.

Signed-off-by: NeilBrown <neil@brown.name>