render-present: render-line-prev fix
When called with '0', we just want to find start of current
line in case anything has changed.  So just go to a new line.
Going to start-of-page can cause confusion here because if already
at start of page (as we normally are) this reports and error
which can confuse render()

Signed-off-by: NeilBrown <neil@brown.name>