Introduce doc:mark-same-exact
If a pane defines an over-ride doc:mark-same to, for example,
hide some elements in the document, it will confuse mark_prev_pane,
mark_next_pane which need to know exactly what is where so that
it can keep all marks in the correct order.

So introduce doc:mark-same-exact which true documents must
implement like doc:mark-same, and which mark_{prev,next}_pane use
to keep marks in the right order.

Panes must not over-ride doc:mark-same-exact

Signed-off-by: NeilBrown <neil@brown.name>