lang-python: don't transparently call python "Commands" from python.
The calling convention for "Commands" in python is very convenient
but a bit odd: args are determined by the type.
If python code tries to call a python "Command" directly it will
use the wrong convention and fail.

There might be a cleaner way to manage this, but for now just
disable the clevers, so the args are translated to a 'ci' and back again.

Signed-off-by: NeilBrown <neil@brown.name>