Change when 'refresh' is called.
Have event:run return when there is nothing to do and
have it called in a loop.
Then the loop calls 'refresh' on every iteration.  This will
no-op if nothing is damaged of course.

So we don't need anyone else to call refresh.

If something gets damaged, then at the next idle moment
a refresh will happen.

Signed-off-by: NeilBrown <neil@brown.name>