Invent new approach for highlighting specific text.
If a "render:*" attribute is set on a character, then
"map-attr" is sent with mark, attr, value and a callback.
The callback can specify attributes and how many characters to
apply it to.

emacs search is changed to set a "render:search" attribute on the start
of the found string, with the length.  That highlights the whole match.

The are various things that still need fixing:
 - I want to be able to insert a string at the attribute
 - I want "render:X" to work on marks as well
 - I want to call something at start of line to trigger off
   and mark *before* the line which still is in effect.

But this is a good step in the right direction.

Signed-off-by: NeilBrown <neil@brown.name>