render-lines: ensure cursor is always fully displayed.
If cursor moves to a line which is partially displayed, we need to
redraw and make sure the cursor line appears properly.

If the cursor line is higher than the screen, we allow it as long as
the top of line is top of screen.  This might need to be re-visited if
it ever becomes an issue.

Signed-off-by: NeilBrown <neil@brown.name>