Adjustments to pane_refresh
I want to be able to call pane_refresh on sub-branches, so it is
wrong to always set abs_z to 0.

'refresh' can sometimes cause damage, so loop a few times just to be sure.
No more than 3 though.

Remove some current causes of damage-during-refresh which are noise
rather than signal.

Signed-off-by: NeilBrown <neil@brown.name>