Move stepping-over-marks into the doc:step functions
This does result in some code duplication, but whenever a mark
changes reference we really should move with respect to other marks
too, and the doc manager is in the best position to do this efficiently.

This remove the need for the "doc:mark-same-exact" interface, and makes
it easier for other interfaces that move a mark to know what to do.

Signed-off-by: NeilBrown <neil@brown.name>