use comm_call_pane() consistently.
Use comm_call_pane when using the ->handle for a specific pane.
Only use comm_call for ->comm2 calls.

A important change here is to use comm_call_pane rather than call5
when passing mark to a document.  If the doc doesn't handle the call, the parent
will misunderstand the mark.

Signed-off-by: NeilBrown <neil@brown.name>