Change how "scale" is managed.
A display can report a "scale:M" attribute as XxY
X is the pixel width of 'M' in the default font, nominally 10pt
Y is the pixel height.

A pane can get a "scale" attribute which is either:
  A scale factor time 1000, so 1000 is the default 1500 is 50% larger
or
  XxY, meaning adjust scale to fit X 'M's across and Y 'M's down.
  (x:Y,y:Y is also possible with a Y scaled differently for historical reasons).

The function pane_scale(p) or Pane.scale() will report a pair of
numbers being the number of pixels high/wide that an M is in the
current scale.  The 'x' number can be passed to text:size and
draw:text to ensure correct scaling of text.

Signed-off-by: NeilBrown <neil@brown.name>