tile: change tile_split to keep root stable.
We previously increased height of tile tree by adding a new root.
This is a problem if the parent want to keep track of its child.
It is also a problem because when we pane_register() in that parent
it might get confused.

So create a new child and move all other children into it.

Signed-off-by: NeilBrown <neil@brown.name>