notmuch: rearrange 'select' functionality.
The new 'master-view' now takes a larger role in handling
'select' with the per-pane viewers just doing the minimum
that must be done there.

The 'query' documents are not children of the main document,
and don't get handled by 'docs'.  This protects them from being
accessed outside the notmuch application, and will make it easier
to reuse existing documents.

Signed-off-by: NeilBrown <neil@brown.name>