Change get-attr on documents to walk down a stack of pane.
The 'document' might be implemented as a stack of panes.
In that case we need to use pane_attr_get() to walk the stack looking
for an attribute value.

Also we need panes that implement a final document (rather than a filter) to
terminate the search by returning '1' to get-attr requests.

Signed-off-by: NeilBrown <neil@brown.name>