doc: use key_handle to pass doc: commands to documents.
When we find a linkage to a document, only send doc: commands
to it, and use key_handle() so a stack of panes can respond.

We limit to doc: to avoid any handling that the root pane might
impose.
Maybe documents should be stacked on something else .. but then
again, I don't want to waste time sending non-doc: commands down
documents.

Signed-off-by: NeilBrown <neil@brown.name>