Split text renderline into a separate pane.
Instead of being part of doc-text, make the conversion
from characters to lines be a separate pane.
This will allow it to be stacked over synthetic text documents.

This introduces a "text" render target which adds renderline and the
render-lines.

This does not yet support sensible handling of enormous lines.  That
will come next.

Signed-off-by: NeilBrown <neil@brown.name>