Preliminary 'email' document handler.
An "email" document is a catenation of other document views,
each of which is an interpretation of some part of the original
email file.

Currently the "email" document only handles a single part, and
that is an rfc822-header view of the header-section of the email.

Lots of this will get replaced, but it is a starting-point to build on.

Signed-off-by: NeilBrown <neil@brown.name>