render-lines: add comments, tidy up code.
I'm wanting to revisit this code, and it is hard to understand :-(

Most of this change is obvious, except removal of
-               if (!wrap && ret == WRAP && line == start)
-                       ret = 0;

if line == start, then draw_some() will have returned 0.

Other change is that we skip over attributes when truncating a line.
There can be no value in parsing them.

Signed-off-by: NeilBrown <neil@brown.name>