python: mark_dealloc must be more careful.
If not self->owned and mark->mtype != &MarkType, it may not
be safe to derefernce the mark - it might have been freed already.

This happens when a document-view is closed and the point is freed.
That point might be an arg to a python function that hasn't completed yet.

So keep track, in the Mark, if mtype is &MarkType.

Signed-off-by: NeilBrown <neil@brown.name>