lib-crop: allow initial mark to about out-of-range.
Failing a command if a mark is out-of-range is not
a good thing.
This particularly causes problems if you create a new
mark and expect it to be at the start of the doc.
So just crop both marks at the start.
Only fail if the more is moved out of the region.

Signed-off-by: NeilBrown <neil@brown.name>