notmuch: send notification when tags change.
notmuch_list can listen for these and updates its cached
information so display changes prompt.

For this we need doc:notmuch:byid:tags, which is a bit
ugly but might be generlized one day.  We need it because
the notmuch_list cannot open the database itself is it is locked.

Now notmuch_list doesn't need to handle
 "doc:notmuch:remove-tag-*" itself.  That was a worse hack.

Signed-off-by: NeilBrown <neil@brown.name>