notmuch: handle empty threads better.
If a thread appear empty, we need to cope correctly
and not try to display the contents.
While this could happen for real if thread was deleted
from the database, a more likely situation is that no
message match the search any more (e.g. they have been archived)
In that case, try to just find the thread, ignoring other search
details.

Signed-off-by: NeilBrown <neil@brown.name>