lib-libevent - work around bug in libevent
If I regiester an event with libevent to read on
a file descriptor, and then register another event to
read on the same file descriptor before deleting the
first event, then when the first event is deleted,
libevent stops listening on that filedescriptor.

This can happen if the handler for a read closes the file,
opens a new file which gets the same fd, and then asks for
events on that new file.
As lib-libevent doesn't delete the event until after the
callback has run, this results in the old event being removed
after the new event is added, which causes the bug.

So work around this but tracking which fd each event is for,
and when we register a handler for a given fd, check if any
other handler for the fd is already registered, and
unregister it early.

Signed-off-by: NeilBrown <neil@brown.name>