email: refactor lib-rfc822header
lib-rfc822header no longer managed to text buffer full of
headers, it just parses and give the header to other code to
manage.
doc-email now calls lib-rfc822header asking it to put certain headers
in a doc that it can include.

This will make it possible to re-use lib-rfc822header to parse
the headers in multipart parts.

Signed-off-by: NeilBrown <neil@brown.name>