doc-multipart: allow there to be mark after last component.
Currently any mark must be in some document.
Sometimes it is useful to have after all documents.
That was a mark can be used to add a document
at the end of the list, but before the mark.

This requires a few changes, but there is extra cleanup in
here.

Signed-off-by: NeilBrown <neil@brown.name>