doc-docs: add new pane as parent of all the documents.
Having the *Documents* document be the parent of other
documents turned out to be problematic as commands for
one document could fall through to the other and
cause confusion.

It is much easier to just have a separate pane, attached to
the root, which groups the listed documents together.
This allows the *Documents* document to easily be part of that
list.

Signed-off-by: NeilBrown <neil@brown.name>