text: fix text_update_X_after_change
If two marks point to the same place but one
is the end of a chunk and the other is the start
of the next chunk, then one might need updating
while the other doesn't.
We were aborting on the other, without looking at the
one.
So if that case is possible, keep looking.

Signed-off-by: NeilBrown <neil@brown.name>