display-pygtk: attach pixmap to pane nearest root.
We currently attach a pixmap to the pane which is the focus
for a drawing command.   That is normally a leaf and so might
be temporary.

It is more stable and sensible to use a pane closer to the root,
providing it is the right size with no offset.

So do that.

Signed-off-by: NeilBrown <neil@brown.name>