render-complete: use pane_call() instead of comm_call().
These two calls are calling a function on the current pane.
Using comm_call() does avoid a lookup, but this is the only
place where a 'home' is passed to comm_call(), and it will soon
be convenient to remove that possibility.
In general, comm_call() will only be for callback, which don't require
a separate home - in fact, focus==home is all that make sense for callbacks.

So change these two pane_call().

Signed-off-by: NeilBrown <neil@brown.name>