reorder final arguments of pane_call()
'comm2' is special in that there is a call_comm() function
which gives comm2 any early spot in the call() arg list.
So it makes sense for the default spot to be last.
With the new call_ret() calls, and with call_comm(), comm2
in a late position will hardly every be used, which
further justifies making it the final position.

So start moving comm2 to the last position, starting with pane_call()

Signed-off-by: NeilBrown <neil@brown.name>