Revise the call() macros (again).
There is now one main varargs macro, CALL().

CALL(ret, type, target,
     key, comm2a, focus,
     num, mark, str, num2, mark2, str2, x, y, comm2);

'ret' is a symbol identifying a return type.  It is used to choose
a function to call, e.g. do_call_val() or do_call_str() or do_call_pane() etc

'type' is  symbol idenfitying what is being called,
  focus (most common), home, pane, comm.
  This selects whether 'target' is a pane or a comm.

comm2 and comm2a are the same arg. This allow it to appear at different
positions in the arg list.

CALL() is not meant to be called directly, though it can be.
There are
  call()
  comm_call()
  pane_call()
  home_call()
  call_ret()
  call_comm()
  call_pane()

which wrap CALL() in various ways.  More might be added.

Signed-off-by: NeilBrown <neil@brown.name>