Add basic "interactive mark" support.
A doc_data can now hold a 'mark' as well as the 'point'.
Note that this is per-view, not per-document.

"Move-to" can move to delete the mark, as well as moving the point,
and "doc:point" returns both.

Emacs mode can leave a mark, swap point and mark, or delete text between
the two.  Abort clears the mark.

Signed-off-by: NeilBrown <neil@brown.name>