search: use Move-Word or Move-Char to find hand C-w and C-c
C-w and C-c include the following word or char in the search string.
Use Move-Word or Move-Char to find out what that means, so the
definition of 'word' is consistent.

Also don't move esi->end forward, as the search triggered by
Notify:doc:Replace will do that.
Just incase something goes wrong with the search and ->end isn't
moved, limit the number of chars added to 1000.

Signed-off-by: NeilBrown <neil@brown.name>