multipart: normalize marks and provide multipart:part-num attr
multipart marks now are normalized: they never point to
the end of a part, but instead move to the start of the next
part.
This means that an empty part can never be indexed by a mark.

Also add multipart:part-num attribute so we can detect
which part we are in.

Signed-off-by: NeilBrown <neil@brown.name>