notmuch:list: massive rewrite
This changes the nothing search view (list of threads)
to control the view in the view pane, not in the documents.
Mark refs are normalized and the view uses
doc:step-matched and doc:step-thread to move around
the document.

It mostly works, but there seem to be a few bugs.
It is close enough to commit..

Signed-off-by: NeilBrown <neil@brown.name>