multipart: make marks easy to compare for equality.
A multipart mark holds a ref to a mark in the underlying
document.
I want equality comparisons to be easy, so when two marks
point the same place, they both use the same underlying
mark.  This requires refcounting.

This code assumes that all other documents already work this way.
Doing that made it a lot easier, and that is unofficially true already.

Signed-off-by: NeilBrown <neil@brown.name>