doc: support enter/leave in expression movement
This supports something like M-C-d which goes down into
an expression, and M-C-u which steps up out of one.
It isn't quite like emacs as it only steps if it finds
the start or end.
I'm not sure this is really useful, but I don't use these
functions much myself.  I'll wait and see.

Signed-off-by: NeilBrown <neil@brown.name>