emacs: don't pass interactive command name in 'str' arg.
The interactive command name can already be found in the 'key'
so it isn't needed in the 'str' arg.
Leaving 'str' clear leave room for other uses of 'str'.
e.g. lib-popup can be told to re-call the command with
a newly ready value as the string.
The command can then pop-up a pane if no str was given.

Signed-off-by: NeilBrown <neil@brown.name>