search: don't extend search string when no match.
C-w and C-c can extend the search string from the current match.
This doesn't make sense if there isn't a current match, so
don't do anything in that case.

When the search string is empty we still want this to work, so we must
make an empty search string to always match (which seems reasonable).

Signed-off-by: NeilBrown <neil@brown.name>