Add server mode.
This is quite preliminary, but core functionality works.

If lib-server.py is run outside of edlib, it requires a filename,
and sends it to edlib for editing - then waits for a "done" indication.

emacs inteprets C-x-# as sending that "done" indication.
If the documeent hasn't been written, the "done" fails.
If the document has no waiters, choose a document which
does, and visit that.

Signed-off-by: NeilBrown <neil@brown.name>