key-lookup: pass down a simple key to commands.
When a key comes from input, it can have multiple names,
  Enter C-Chr-M
etc.
We need to command to see only the key that was matched, as
it sometimes acts on what it expects to see.
So keep track of what was matches, and send that.

This is a bit of a hack..

Signed-off-by: NeilBrown <neil@brown.name>