search: backspace now deletes auto-inserted text.
If I type C-w to include the rest of the word in the search
string, then Backspace will delete everything that was inserted.

This is managed by setting an attribute on ever extra character
added.
Unfortunately, doc-text has a bug which prevents this from work
for more than the most recent insert.

Signed-off-by: NeilBrown <neil@brown.name>