Add home_call_pane() interface.
This is like home_call() in that it take an extra pane argument
as use that  for 'home' (Rather and duplicating focus).
It is like call_pane() in that it adds a comm2 function to return
a pane from the function.
Hence: home_call_pane() - pass in  home, and return a pane.

Signed-off-by: NeilBrown <neil@brown.name>