separate doc:assign-view from doc:assign
doc:assign currently attaches a document to a gateway AND
adds some viewing frames.
Sometimes we want the one without the other, and have a hack
with 'str==NULL' means don't do the other.

Instead, have doc:assign which just does the assignment,
and doc:assign-view which does the assignment and adds viewers.
That is what is normally wanted.

Signed-off-by: NeilBrown <neil@brown.name>