separate "render" from "view" config.
When displaying a document, we always need a "render" pane
to provide "Refresh" support, and sometime need a "view"
pane to adjusts the style of the view - C mode or viewer
mode or similar.

Separate these out more clearly and use the two strings passed
to doc:assign-view to describe each, rather than trying to
encode both into the one string.
Now there is less confusion between a 'render' pane and a
'view' pane.

Signed-off-by: NeilBrown <neil@brown.name>