grep: add git-grep support.
This adds support to emacs to have "simple" commands
which always pass a numeric arg.
If a grep command gets that, it find a word in doc and suggests
that as a default.

A negative arg to git-grep does the grep in current directory,
else we walk up to find the top-level directory.

M-.  does a git-grep
M--M. does it in local directory

Signed-off-by: NeilBrown <neil@brown.name>