doc-text: fix (I hope) undo problems.
Keeping a redo history after making a change
is broken, as the redo edits are based on a different
document to what we have.
Instead, the redo history must be attached to
the relavant point in the undo history.

So do that, and allow 'undo' to walk around the
complete tree.  So we don't have an 'redo' function
any more.

Signed-off-by: NeilBrown <neil@brown.name>