render-lines: record width of cursor
When checking if the cursor fit on the page, to decide if side-scrolling
is needed, we need to know the cursor width too.
So assume it if mwidth and pass it back.  Then use it to decide
if scrolling is needed.

Without this, side-scroll only works when the pane is an exact number
of characters wide.

Signed-off-by: NeilBrown <neil@brown.name>