ncurses: hack to work-around ncurses bug.
ncurses keeps all "windows" of all "screen" in a linked list.
When you call resize_term on a given screen, it acts of
*all* windows on the list.  Ouch.

So I need to keep the lists for the multiple screens separate.
I "know" that stdscr is  always the first window, so I find the
offset in _nc_globals where that is stored, and:
  - clear it before creating a new window
  - restore there before doing anything else to a window

Signed-off-by: NeilBrown <neil@brown.name>