grep: add more options to M-. "git-grep".
Check the parents to see if this really is a git tree (.git
exists) or maybe a quilt tree (.pc exists) or if there is
a TAGS file.

I'd like to extra filenames from the TAGS file one day,
but today is not that day.

If a marker is found, grep or git grep is run there,
otherwise it is run in current directory.

Signed-off-by: NeilBrown <neil@brown.name>