mark: replace 'state' with 'owner' in a view
Each view on a doc now has an 'owner' which is
a pane.  This subsume the role of 'state' which
was either 0 or 1 (now owner is NULL or non-NULL) -
2 hasn't been used for a while.

When doc:add-view is called, the focus must be the
pane that will own the view, and the same pane must
be passed to doc:del-view.

Signed-off-by: NeilBrown <neil@brown.name>