doc-text: fix adjustments after undo.
In particular, text_retreat_towards() didn't do what the
spec says.  If in the middle of a chunk it would move
to the end of the previous rather than the start of this.
That messed up the repositioning of point.

Also improve the code for finding the edit point, so
that it tries harder if a close match cannot be found.

This should fix a crash that happens during undo very
occasionally.

Signed-off-by: NeilBrown <neil@brown.name>