pane: enhance reparent, and add python binding.
Currently, reparent requires the new parent to be
a sibling.  Allow it to also be a parent-less pane,
which is inserted as a child of the previous parent.

This is useful for reparenting a pane underneath the tiler
as it needs to only have one child at a time.

When a parentless pane is introduced as a new parent,
set set the size to match its new parent.

Also inform the original parent when child is moved or
replaced.

Signed-off-by: NeilBrown <neil@brown.name>