notmuch: discard 'maindoc'.
We are storing in 'self.maindoc' exactly the pane that
is self.parent.  This is pointless.
Also, we rarely need to access that pane directly -
just once.  Other case we get is implicily in the same
way that any view pane accesses the document.
So discard it, and no need to close when it closes,
that happens automatically.

Signed-off-by: NeilBrown <neil@brown.name>