doc: discard doc:set-parent in favour of reparent
We don't need doc:set-parent as whenever we want to reparent
a doc, we have the actual doc pane and can act on it directly.
So change to use pane_reparent().

email messages were being reparented in an unhelpful way,
so discard that for now.  When I get back to working on
notmuch mode, I can work out what is required there, if anything.

Also fix a bug in doc-docs where the docs collection was reparented
to itself, rather than the doc-docs document being reparented there.

Signed-off-by: NeilBrown <neil@brown.name>