doc: change mechanism for attaching document in a view pane
Instead of two separate commands we now have a single command
sent to the document, with the attachment pane as the 'focus'.
The default implementation does all that the previous code did.

The benefit - other than simplicity - is that a document has
more control and can build arbitrarily complex stacks.

Signed-off-by: NeilBrown <neil@brown.name>