Python: fix refcounting on panes.
The ->handle holds the counted reference on Panes and Docs,
and it is released (command_put) when the pane is closed.
We currently take one too many references, so it never gets
closed.
A consequence of this is that bracket-highlight marks aren't
destroyed when a pane is closed, so they hang around.

We previously had explicitly code in the "Close" handler for Doc,
but not for Pane, to drop the reference.  This has now been removed
before the command.free handler does everything.

Signed-off-by: NeilBrown <neil@brown.name>