text: check when file changes.
If the file has changed when we are about to
modify an un-modified document, switch to read-only.
On switch back to read/write, allow changes.

Display the 'file changed' status in the status line.

Clear 'file changed' status when we load or save the file.

When auto-saving, check if the file has changed, and if so,
update the status.

Signed-off-by: NeilBrown <neil@brown.name>