make: don't ask MakePane to open the view
When we send a notification via a document, the recipient
won't know which pane sent the notification - it will (soon)
seem come from the document.
So in response to "doc:make-revisit", the MakePane cannot
open the document - it doesn't know where.
It can only move "point".  The caller then asks it
to open the doc with a different sort of notification.

.. there is room for improvement here...

Signed-off-by: NeilBrown <neil@brown.name>