Remove 'rpos' functionality.
This was a design mistake and is going away.

The doc-rendering pane can do something similar, and in
some ways much better.
If something more similar is needed, the 'render pos' should
be stored in a display pane associated with just the 'point',
not with every mark.  It is then used to display the cursor.
The same pane would capture left/right movement and update
the curspor pos.

Note that this breaks some aspects of email handling.
That needs a lot of work anyway...

Signed-off-by: NeilBrown <neil@brown.name>