editor: support delayed freeing of marks
Much like panes, we need an rcu-like freeing of marks.

A particular problem is that marks in an active command
cannot be NULLed when the pane is freed, unlike for
example the point stored by lib-input which can be.

Most that mark should not be used by anything, but
a support library like python my need to look inside
just in case.
So we clear most content and set type to MARK_UNGROUPED
so support libraries can know to igore the test.

Signed-off-by: NeilBrown <neil@brown.name>