search: handle zero-len matches better.
Normally when we 'search again' we want to start from
where the last match finished.  But if that match was
zero characters, that doesn't get us anywhere - we need
to move forward one char, or to other end of file
if we had reached the end.

The previous approach meant that the initial search would
sometimes not match at the start point.

Signed-off-by: NeilBrown <neil@brown.name>