rexel: restructure, and add '\b' support.
For '\b' I needed some more "special" rexels, so I
documented what I have a bit more clearly and rearranged
the number assignment.
Now number for 'sets' have been reduced a bit, and we
have 32 numbers available for "special".

I've also change whether '.' matches a newline to be
a different sort of special, rather than a global flag.

I've added a flag for 'lax' matching, but not implemented
it yet.
I've also added a special for '\B' - not a word break.
But that isn't implemented either because the current flag
handling doesn't support 'not' flags.

Bit '\b' works - it matches either start of end of a word.

Signed-off-by: NeilBrown <neil@brown.name>

# Conflicts:
#	rexel.c