Revise how docs are postioned in the docs list.
When they first appear they go to the end of the list,
unless a positive number is given to doc-appeared-

doc:attach-view leaves it where it is unless a positive or
negative number is given.

Most user-requested visit of docs give a positive number.
Some visits of uninteresting docs give a negative numbers.

Signed-off-by: NeilBrown <neil@brown.name>