Central handling of DAMAGED_CURSOR
Also, fix render-lines so that it tracks whether the
cursor has moved better.  This is needed for target_x
and ignore_cursor handling.
The best way is to leverage the fact that we already keep
a copy of the previous cursor.  If that moved and we
didn't move it (preserving target_x), then reset
so that we track where the cursor actually is.

Signed-off-by: NeilBrown <neil@brown.name>