rexel: fix set_match for inverted char classes
If we inverted a set of multiple char classes, we always
matches, because a target will fail to match one of the classes.
The condition was messed up. As soon as we find a match we must
return, either success or failure.

Also fix up test infrastructure for testing for failure-to-match.

Signed-off-by: NeilBrown <neil@brown.name>